+++ src/main/java/bc1/gream/domain/product/controller/ProductQueryController.java
@@ -3,10 +3,10 @@ package bc1.gream.domain.product.controller;
 import bc1.gream.domain.common.facade.BuyOrderQueryFacade;
 import bc1.gream.domain.common.facade.ProductOrderQueryFacade;
 import bc1.gream.domain.common.facade.SellOrderQueryFacade;
-import bc1.gream.domain.order.entity.Buy;
+import bc1.gream.domain.buy.entity.Buy;
 import bc1.gream.domain.order.entity.Order;
 import bc1.gream.domain.order.entity.Sell;
-import bc1.gream.domain.order.mapper.BuyMapper;
+import bc1.gream.domain.buy.mapper.BuyMapper;
 import bc1.gream.domain.order.mapper.OrderMapper;
 import bc1.gream.domain.order.mapper.SellMapper;
 import bc1.gream.domain.product.dto.response.ProductQueryResponseDto;
